+++ src/OpenSimCreator/Registry/StaticComponentRegistries.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace osc { template<typename> class ComponentRegistry; }
+
+namespace osc
+{
+    template<typename T>
+    ComponentRegistry<T> const& GetComponentRegistry();
+}